solve training loop
